requirements.txt
# RAG_System - Global Requirements
# Python 3.8+

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & NLP
scikit-learn>=1.3.0
jieba>=0.42.1

# Deep Learning & Transformers (Optional - for BERT)
# transformers>=4.30.0
# torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# LangChain & OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# Vector Store & Database
chromadb>=0.4.0
neo4j>=5.0.0
# Web Framework
streamlit>=1.28.0

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.0

# System Utilities
psutil>=5.9.0

# Development Tools
# jupyter>=1.0.0
# notebook>=7.0.0
